Compute age standardized rate based on provided standard population

This function compute the crude rate and age standard rates and their 95% confidence intervals(CIs).

Usage

asr_rate(
  data,
  ...,
  method = "gamma",
  event = fbs,
  pop = rks,
  agegrp = "agegrp",
  std = c("segi", "china"),
  mp = 1e+05,
  sig = 0.95,
  decimal = 2,
  show_var = FALSE,
  type = "big",
  lang = "cn"
)

Arguments

data

A data frame or tibble that containing case counts, population in risk, and stratification variables for calculating standardized rates.

...

Variable or variables for stratification when compute age standard rates.

method

This parameter specifies the method used to calculate confidence intervals for the rates. The options are "gamma" (default), "normal", or "lognormal". Different methods may produce slightly different confidence intervals.

event

A variable within the input data that corresponds to the cases counts.

pop

A variable within the input data that corresponds to the population in risk in each stratification.

agegrp

A variable within the input data that corresponds to the age categories.

std

Specify the standard population structure in the 'std_pop' data frame used for calculating standardized rates. When calculating standardized rates for multiple standard populations, specify std = c("segi", "china").

mp

A constant to multiply rates by (e.g. mp=1000 for rates per 1000).

sig

The desired level of confidence in computing confidence intervals. The default is 0.95 for 95 percent CIs.

decimal

This parameter specifies the number of decimal places to round the results. The default is 2, which means rates will be rounded to two decimal places.

show_var

Logical value whether show variance or not.

type

Character, regrouping type for cancers, options are 'system', 'big', or 'small', default is 'big'.

lang

Character, language for output, options are 'en' or 'cn'. Default is 'cn'.

Value

A data frame or tibble contains the age standard rates and CIs. Variables included in the output data frame as follows.

...

Variable or variables for stratification.

no_cases

Sum of cases in each stratification.

cr

Crude rate in each stratification.

cr_var

Variance of the crude rate.

cr_lower

Lower limit of the crude rate confidence interval.

cr_upper

Upper limit of the crude rate confidence interval.

asr_china

Age standardized rate using the china standard pop. The output variable name would be "asr_" followed by the name of the standard population, if other standard population was used.

asr_var_china

Variance of the age standard rate using china standard population.

asr_lower_china

Lower limit of the age standard rate confidence interval using china standard population.

asr_upper_china

Upper limit of the age standard rate confidence interval using china standard population.

References

Fay MP, Feuer EJ (1997). “Confidence Intervals for Directly Standardized Rates: A Method Based on the Gamma Distribution.” Stat Med, 16(7), 791--801. ISSN 0277-6715, http://www.ncbi.nlm.nih.gov/pubmed/9131766.

Author

Qiong Chen