Convert ICD10 Codes to Cancer Categories.

Convert ICD10 codes to Cancer Categories according to the specified category type and language.

classify_icd10(icd10, type = "big", lang = "cn")

Arguments

icd10

The ICD10 codes of cancer part (C00-C98 and D00-D48) collected by the Population-Based Cancer Registration (PBCR).

type

Character, type of cancer category ("system","big","small"), by default, the "big" category was choosed which will categorize the icd10 code 25 cancer categories.

lang

Character, two options ("cn","en"), by default, lang="cn" was choosed, which will output the categories in Chinese, otherwise, in English.

Value

a vector that contain the cancer categories.

References

National Cancer Center. Guidelines for Cancer Registration in China (2016) [M]. Beijing: People's Medical Publishing House, 2016.

Author

Qiong Chen, Email: chenqiong@hnccr.org.cn

Examples

library(canregtools)
icd10 <- c("C16.1", "C15.1", "C34.1", "C34.2", "C15.0")
category <- classify_icd10(icd10, type = "system", lang = "en")