Clean canreg data.

Clean canreg data.

Usage

clean_canreg(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

# S3 method for class 'FBcases'
clean_canreg(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

# S3 method for class 'SWcases'
clean_canreg(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

# S3 method for class 'population'
clean_canreg(
  x,
  cutage_method = "distance",
  smooth_age = "natural",
  breaks = c(0, 15, 40, 65),
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  labels = NULL,
  label_tail = NULL,
  cancer_type = "big"
)

Arguments

x

Data of class 'FBcases', 'SWcases' or 'population'.

cutage_method

Method for grouping ages. Options are 'distance', 'interval', or 'quarter'. Default is 'distance'.

smooth_age

Method for expanding age. Options are 'linear','natural', or 'periodic'. Default is 'natural'.

breaks

Cut points for age groups when 'cutage_method' is 'interval'. Default is 'c(0, 15, 40, 65)'.

length

Length of each age group when 'cutage_method' is 'distance'. Default is 5.

maxage

Start age of the last age groups. Default is 85.

sep_zero

Logical value TRUE or FALSE, if the 0-year age group treated as a separate group. Default is TRUE.

labels

Labels for the age groups.

label_tail

Tail of the labels.

cancer_type

A character string specifying the classification method used to categorize ICD-10 codes. This determines how ICD-10 codes are classified. Options include "big" (classify ICD-10 codes into 26 cancer categories), "small" (classify ICD-10 codes into 59 cancer categories, more specific categories), "system" (classify ICD-10 codes into organ system), and "gco" (classify ICD-10 code into cancer categories same as classification published by the Global Cancer Observatory). This parameter is only available when the input data is a vector of ICD-10 codes, or object with class of 'canreg' or 'canregs'.

Value

Class 'canreg'.