Estimating population structure using interpolation method.

Estimating population structure using interpolation method.

Usage

estimate_pop(pop1, pop2, period)

Arguments

pop1

Population or population proportion in each age group for the start year.

pop2

Population or population proportion in each age group for the end year.

period

Vector contain the start year and end year value.

Value

A data frame contain the estimated population proportion in each year during the period with each year in one column and each age group in one row.

Examples

pop1 <- c(59546, 294129, 472511, 552549, 821119, 996436, 805635, 1004506,
          989357, 1056612, 986559, 792270, 544544, 452297, 473579, 350802,
          212614, 109598, 61990)
pop2 <- c(75641, 377276, 327116, 380338, 539034, 1158852, 1152329, 881443,
          903484, 1011164, 1238871, 1137832, 1022787, 645441, 464777,
          482941, 406144, 227977, 144526)
estimate_pop(pop1, pop2, c(2000, 2010))
#>           2000        2001        2002        2003        2004        2005
#> 1  0.005395295 0.005457142 0.005518990 0.005580837 0.005642684 0.005704531
#> 2  0.026650199 0.026984677 0.027319155 0.027653632 0.027988110 0.028322588
#> 3  0.042812889 0.041132305 0.039451722 0.037771138 0.036090554 0.034409971
#> 4  0.050064906 0.048082258 0.046099609 0.044116960 0.042134311 0.040151662
#> 5  0.074399277 0.071244888 0.068090500 0.064936112 0.061781724 0.058627336
#> 6  0.090284256 0.090469175 0.090654094 0.090839013 0.091023932 0.091208851
#> 7  0.072996315 0.074858168 0.076720020 0.078581873 0.080443726 0.082305578
#> 8  0.091015456 0.088921740 0.086828025 0.084734310 0.082640594 0.080546879
#> 9  0.089642847 0.087861628 0.086080408 0.084299189 0.082517969 0.080736750
#> 10 0.095736633 0.094202135 0.092667637 0.091133138 0.089598640 0.088064141
#> 11 0.089389328 0.090299924 0.091210519 0.092121115 0.093031710 0.093942305
#> 12 0.071785350 0.073653042 0.075520734 0.077388426 0.079256118 0.081123810
#> 13 0.049339596 0.052537209 0.055734822 0.058932435 0.062130048 0.065327661
#> 14 0.040981355 0.042014738 0.043048121 0.044081504 0.045114887 0.046148270
#> 15 0.042909657 0.042313858 0.041718058 0.041122259 0.040526459 0.039930659
#> 16 0.031785180 0.032446239 0.033107299 0.033768358 0.034429417 0.035090477
#> 17 0.019264355 0.020566929 0.021869504 0.023172078 0.024474653 0.025777227
#> 18 0.009930366 0.010749839 0.011569312 0.012388786 0.013208259 0.014027732
#> 19 0.005616739 0.006204106 0.006791472 0.007378839 0.007966205 0.008553572
#>           2006        2007        2008       2009        2010
#> 1  0.005766378 0.005828225 0.005890073 0.00595192 0.006013767
#> 2  0.028657065 0.028991543 0.029326021 0.02966050 0.029994976
#> 3  0.032729387 0.031048803 0.029368220 0.02768764 0.026007052
#> 4  0.038169013 0.036186364 0.034203715 0.03222107 0.030238418
#> 5  0.055472948 0.052318559 0.049164171 0.04600978 0.042855395
#> 6  0.091393770 0.091578689 0.091763608 0.09194853 0.092133446
#> 7  0.084167431 0.086029283 0.087891136 0.08975299 0.091614841
#> 8  0.078453164 0.076359449 0.074265733 0.07217202 0.070078303
#> 9  0.078955530 0.077174311 0.075393091 0.07361187 0.071830652
#> 10 0.086529643 0.084995145 0.083460646 0.08192615 0.080391650
#> 11 0.094852901 0.095763496 0.096674091 0.09758469 0.098495282
#> 12 0.082991503 0.084859195 0.086726887 0.08859458 0.090462271
#> 13 0.068525274 0.071722887 0.074920500 0.07811811 0.081315726
#> 14 0.047181653 0.048215036 0.049248418 0.05028180 0.051315184
#> 15 0.039334860 0.038739060 0.038143261 0.03754746 0.036951661
#> 16 0.035751536 0.036412595 0.037073655 0.03773471 0.038395773
#> 17 0.027079802 0.028382376 0.029684951 0.03098753 0.032290100
#> 18 0.014847205 0.015666679 0.016486152 0.01730563 0.018125099
#> 19 0.009140938 0.009728305 0.010315672 0.01090304 0.011490405