Convert ICDO3 codes to ICD10 code.

Convert the ICDO3 codes to ICD10 code based on Topography, Morphology, and Behavior value.

icdo3to10(topo, morp, beha)

Arguments

topo

A vector containing ICDO3 topography codes, it can be in the format of either "C15.1" or "C151".

morp

A vector containing ICDO3 morphology codes, it could be in the format of "M8170" or "8170".

beha

A vector containing ICDO3 behavior codes, it could be one of the values c("0", "1", "2", "3", "6").

Value

A vector containing ICD10 code.

Details

ICD-O-3 (International Classification of Diseases for Oncology, 3rd Edition) and ICD-10 (International Classification of Diseases, 10th Edition) are two different coding systems used for disease classification. ICD-10 is a widely used coding system that can describe various diseases, including tumor diseases. On the other hand, ICD-O-3 is specifically designed for the coding of tumor diseases. Promoting and using ICD-O-3 and ICD-10 helps in standardizing and unifying the classification and coding of tumor, facilitating global tumor epidemiology research, statistical analysis, and clinical management. ICD-O-3 consists of three main components: Topography, Morphology, and Behavior. The Topography code uses a four-digit number to describe the specific location where the tumor occurs. The Morphology code uses a three-digit number to represent the morphology or histology type of the tumor. The Behavior code describes the behavior characteristics of the tumor. With this program, you can convert ICD-O-3 codes to ICD-10 codes, making it convenient for research and clinical practice.

Examples

topo <- c("C001", "C15.1", "C349", "509", "C421", "C774")
morp <- c("8140", "8000", "8070", "8070", "8000", "8070")
beha <- rep("3", 6)