Read cancer registration data in the format reported to the National Cancer Center.

Read cancer registration data in the format reported to the National Cancer Center.

Usage

read_canreg(x, pop_type = "long")

Arguments

x

Address of cancer registration data, it could be a specific excel file, or a folder contains a series of Excel files.

pop_type

Format of population sheet, options area 'long' or 'wide', default is 'long'.

Value

An object of canreg class or a list of objects of canreg class.

Examples

library(canregtools)
file <- system.file("extdata", "411721.xls", package = "canregtools")
data <- read_canreg(file)