Count canreg data.

Count canreg data.

count_canreg(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

# S3 method for class 'canregs'
count_canreg(x, ...)

# S3 method for class 'canreg'
count_canreg(
  x,
  age_breaks = c(0, 1, seq(5, 85, 5)),
  label_tail = NULL,
  cancer_type = "big"
)

Arguments

x

Object data of class 'canreg' or 'canregs'.

age_breaks

Specify the break points classify age groups. Default is c(0, 1, seq(0, 85, 5)).

label_tail

Tail label to be added to the labels. Default is NULL.

cancer_type

A character string specifying the classification method used to categorize ICD-10 codes. This determines how ICD-10 codes are classified. Options include "big" (classify ICD-10 codes into 26 cancer categories), "small" (classify ICD-10 codes into 59 cancer categories, more specific categories), "system" (classify ICD-10 codes into organ system), and "gco" (classify ICD-10 code into cancer categories same as classification published by the Global Cancer Observatory). This parameter is only available when the input data is a vector of ICD-10 codes, or object with class of 'canreg' or 'canregs'.

...

Parameters.

Value

data of class fbswicd.

Examples

library(canregtools)
data("canregs")
fbsw <- count_canreg(canregs)
fbsw
#> $`410102`
#> $areacode
#> [1] "410102"
#> 
#> $fbswicd
#> # A tibble: 1,596 × 11
#>     year   sex cancer agegrp     fbs   sws    mv   dco m8000    ub   sub
#>    <int> <int> <chr>  <fct>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2021     0 101    0 岁         0     0     0     0     0     0     0
#>  2  2021     0 101    1-4 岁       0     0     0     0     0     0     0
#>  3  2021     0 101    5-9 岁       0     0     0     0     0     0     0
#>  4  2021     0 101    10-14 岁     0     0     0     0     0     0     0
#>  5  2021     0 101    15-19 岁     0     0     0     0     0     0     0
#>  6  2021     0 101    20-24 岁     0     0     0     0     0     0     0
#>  7  2021     0 101    25-29 岁     0     0     0     0     0     0     0
#>  8  2021     0 101    30-34 岁     1     0     1     0     0     0     1
#>  9  2021     0 101    35-39 岁     0     0     0     0     0     0     0
#> 10  2021     0 101    40-44 岁     1     0     1     0     0     0     0
#> # ℹ 1,586 more rows
#> 
#> $sitemorp
#> # A tibble: 45 × 5
#>     year   sex cancer site       morp      
#>    <int> <int> <chr>  <list>     <list>    
#>  1  2021     1 101    <int [13]> <int [4]> 
#>  2  2021     1 102    <int [1]>  <int [2]> 
#>  3  2021     1 103    <int [8]>  <int [5]> 
#>  4  2021     1 104    <int [8]>  <int [9]> 
#>  5  2021     1 105    <int [12]> <int [9]> 
#>  6  2021     1 106    <int [5]>  <int [9]> 
#>  7  2021     1 107    <int [5]>  <int [6]> 
#>  8  2021     1 108    <int [4]>  <int [7]> 
#>  9  2021     1 109    <int [3]>  <int [5]> 
#> 10  2021     1 110    <int [7]>  <int [19]>
#> # ℹ 35 more rows
#> 
#> $pop
#> # A tibble: 38 × 4
#>     year   sex agegrp     rks
#>    <int> <int> <fct>    <dbl>
#>  1  2021     1 0 岁      3850
#>  2  2021     1 1-4 岁   14405
#>  3  2021     1 5-9 岁   15176
#>  4  2021     1 10-14 岁 16227
#>  5  2021     1 15-19 岁 18685
#>  6  2021     1 20-24 岁 27799
#>  7  2021     1 25-29 岁 36206
#>  8  2021     1 30-34 岁 29301
#>  9  2021     1 35-39 岁 29868
#> 10  2021     1 40-44 岁 30616
#> # ℹ 28 more rows
#> 
#> attr(,"class")
#> [1] "fbswicd" "list"   
#> 
#> $`410103`
#> $areacode
#> [1] "410103"
#> 
#> $fbswicd
#> # A tibble: 1,596 × 11
#>     year   sex cancer agegrp     fbs   sws    mv   dco m8000    ub   sub
#>    <int> <int> <chr>  <fct>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2021     0 101    0 岁         0     0     0     0     0     0     0
#>  2  2021     0 101    1-4 岁       0     0     0     0     0     0     0
#>  3  2021     0 101    5-9 岁       0     0     0     0     0     0     0
#>  4  2021     0 101    10-14 岁     0     0     0     0     0     0     0
#>  5  2021     0 101    15-19 岁     1     0     1     0     0     0     0
#>  6  2021     0 101    20-24 岁     1     0     1     0     0     0     1
#>  7  2021     0 101    25-29 岁     0     0     0     0     0     0     0
#>  8  2021     0 101    30-34 岁     1     0     1     0     0     0     0
#>  9  2021     0 101    35-39 岁     0     0     0     0     0     0     0
#> 10  2021     0 101    40-44 岁     2     0     2     0     0     0     0
#> # ℹ 1,586 more rows
#> 
#> $sitemorp
#> # A tibble: 45 × 5
#>     year   sex cancer site       morp      
#>    <int> <int> <chr>  <list>     <list>    
#>  1  2021     1 101    <int [14]> <int [5]> 
#>  2  2021     1 102    <int [1]>  <int [2]> 
#>  3  2021     1 103    <int [7]>  <int [6]> 
#>  4  2021     1 104    <int [7]>  <int [8]> 
#>  5  2021     1 105    <int [14]> <int [12]>
#>  6  2021     1 106    <int [5]>  <int [7]> 
#>  7  2021     1 107    <int [4]>  <int [4]> 
#>  8  2021     1 108    <int [5]>  <int [6]> 
#>  9  2021     1 109    <int [2]>  <int [3]> 
#> 10  2021     1 110    <int [7]>  <int [19]>
#> # ℹ 35 more rows
#> 
#> $pop
#> # A tibble: 38 × 4
#>     year   sex agegrp     rks
#>    <int> <int> <fct>    <dbl>
#>  1  2021     1 0 岁      3575
#>  2  2021     1 1-4 岁   19832
#>  3  2021     1 5-9 岁   28744
#>  4  2021     1 10-14 岁 21277
#>  5  2021     1 15-19 岁 16109
#>  6  2021     1 20-24 岁 15393
#>  7  2021     1 25-29 岁 18171
#>  8  2021     1 30-34 岁 28362
#>  9  2021     1 35-39 岁 24401
#> 10  2021     1 40-44 岁 22083
#> # ℹ 28 more rows
#> 
#> attr(,"class")
#> [1] "fbswicd" "list"   
#> 
#> $`410307`
#> $areacode
#> [1] "410307"
#> 
#> $fbswicd
#> # A tibble: 1,596 × 11
#>     year   sex cancer agegrp     fbs   sws    mv   dco m8000    ub   sub
#>    <int> <int> <chr>  <fct>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2021     0 101    0 岁         0     0     0     0     0     0     0
#>  2  2021     0 101    1-4 岁       0     0     0     0     0     0     0
#>  3  2021     0 101    5-9 岁       0     0     0     0     0     0     0
#>  4  2021     0 101    10-14 岁     0     0     0     0     0     0     0
#>  5  2021     0 101    15-19 岁     0     0     0     0     0     0     0
#>  6  2021     0 101    20-24 岁     0     0     0     0     0     0     0
#>  7  2021     0 101    25-29 岁     0     0     0     0     0     0     0
#>  8  2021     0 101    30-34 岁     1     0     1     0     0     0     0
#>  9  2021     0 101    35-39 岁     0     0     0     0     0     0     0
#> 10  2021     0 101    40-44 岁     0     0     0     0     0     0     0
#> # ℹ 1,586 more rows
#> 
#> $sitemorp
#> # A tibble: 45 × 5
#>     year   sex cancer site       morp      
#>    <int> <int> <chr>  <list>     <list>    
#>  1  2021     1 101    <int [14]> <int [3]> 
#>  2  2021     1 102    <int [1]>  <int [1]> 
#>  3  2021     1 103    <int [5]>  <int [6]> 
#>  4  2021     1 104    <int [8]>  <int [7]> 
#>  5  2021     1 105    <int [10]> <int [5]> 
#>  6  2021     1 106    <int [4]>  <int [8]> 
#>  7  2021     1 107    <int [5]>  <int [4]> 
#>  8  2021     1 108    <int [4]>  <int [3]> 
#>  9  2021     1 109    <int [2]>  <int [2]> 
#> 10  2021     1 110    <int [5]>  <int [11]>
#> # ℹ 35 more rows
#> 
#> $pop
#> # A tibble: 38 × 4
#>     year   sex agegrp     rks
#>    <int> <int> <fct>    <dbl>
#>  1  2021     1 0 岁      2040
#>  2  2021     1 1-4 岁   16184
#>  3  2021     1 5-9 岁   17569
#>  4  2021     1 10-14 岁 18286
#>  5  2021     1 15-19 岁 14254
#>  6  2021     1 20-24 岁 10963
#>  7  2021     1 25-29 岁 11106
#>  8  2021     1 30-34 岁 20840
#>  9  2021     1 35-39 岁 21487
#> 10  2021     1 40-44 岁 16946
#> # ℹ 28 more rows
#> 
#> attr(,"class")
#> [1] "fbswicd" "list"   
#> 
#> $`410581`
#> $areacode
#> [1] "410581"
#> 
#> $fbswicd
#> # A tibble: 1,596 × 11
#>     year   sex cancer agegrp     fbs   sws    mv   dco m8000    ub   sub
#>    <int> <int> <chr>  <fct>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2021     0 101    0 岁         0     0     0     0     0     0     0
#>  2  2021     0 101    1-4 岁       0     0     0     0     0     0     0
#>  3  2021     0 101    5-9 岁       0     0     0     0     0     0     0
#>  4  2021     0 101    10-14 岁     0     0     0     0     0     0     0
#>  5  2021     0 101    15-19 岁     0     0     0     0     0     0     0
#>  6  2021     0 101    20-24 岁     0     0     0     0     0     0     0
#>  7  2021     0 101    25-29 岁     0     0     0     0     0     0     0
#>  8  2021     0 101    30-34 岁     0     0     0     0     0     0     0
#>  9  2021     0 101    35-39 岁     1     0     1     0     0     0     1
#> 10  2021     0 101    40-44 岁     2     0     2     0     0     0     1
#> # ℹ 1,586 more rows
#> 
#> $sitemorp
#> # A tibble: 45 × 5
#>     year   sex cancer site       morp      
#>    <int> <int> <chr>  <list>     <list>    
#>  1  2021     1 101    <int [9]>  <int [3]> 
#>  2  2021     1 102    <int [2]>  <int [1]> 
#>  3  2021     1 103    <int [6]>  <int [6]> 
#>  4  2021     1 104    <int [8]>  <int [11]>
#>  5  2021     1 105    <int [11]> <int [8]> 
#>  6  2021     1 106    <int [4]>  <int [4]> 
#>  7  2021     1 107    <int [5]>  <int [4]> 
#>  8  2021     1 108    <int [2]>  <int [3]> 
#>  9  2021     1 109    <int [3]>  <int [1]> 
#> 10  2021     1 110    <int [6]>  <int [10]>
#> # ℹ 35 more rows
#> 
#> $pop
#> # A tibble: 38 × 4
#>     year   sex agegrp     rks
#>    <int> <int> <fct>    <dbl>
#>  1  2021     1 0 岁      3429
#>  2  2021     1 1-4 岁   20662
#>  3  2021     1 5-9 岁   42881
#>  4  2021     1 10-14 岁 52657
#>  5  2021     1 15-19 岁 44397
#>  6  2021     1 20-24 岁 29275
#>  7  2021     1 25-29 岁 28104
#>  8  2021     1 30-34 岁 62070
#>  9  2021     1 35-39 岁 40492
#> 10  2021     1 40-44 岁 33294
#> # ℹ 28 more rows
#> 
#> attr(,"class")
#> [1] "fbswicd" "list"   
#> 
#> $`411624`
#> $areacode
#> [1] "411624"
#> 
#> $fbswicd
#> # A tibble: 1,596 × 11
#>     year   sex cancer agegrp     fbs   sws    mv   dco m8000    ub   sub
#>    <int> <int> <chr>  <fct>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2021     0 101    0 岁         0     0     0     0     0     0     0
#>  2  2021     0 101    1-4 岁       0     0     0     0     0     0     0
#>  3  2021     0 101    5-9 岁       0     0     0     0     0     0     0
#>  4  2021     0 101    10-14 岁     0     0     0     0     0     0     0
#>  5  2021     0 101    15-19 岁     0     0     0     0     0     0     0
#>  6  2021     0 101    20-24 岁     0     0     0     0     0     0     0
#>  7  2021     0 101    25-29 岁     0     0     0     0     0     0     0
#>  8  2021     0 101    30-34 岁     4     1     4     0     0     0     0
#>  9  2021     0 101    35-39 岁     0     0     0     0     0     0     0
#> 10  2021     0 101    40-44 岁     2     0     1     0     1     0     0
#> # ℹ 1,586 more rows
#> 
#> $sitemorp
#> # A tibble: 46 × 5
#>     year   sex cancer site       morp      
#>    <int> <int> <chr>  <list>     <list>    
#>  1  2021     1 101    <int [15]> <int [7]> 
#>  2  2021     1 102    <int [5]>  <int [4]> 
#>  3  2021     1 103    <int [8]>  <int [6]> 
#>  4  2021     1 104    <int [9]>  <int [9]> 
#>  5  2021     1 105    <int [13]> <int [9]> 
#>  6  2021     1 106    <int [5]>  <int [11]>
#>  7  2021     1 107    <int [3]>  <int [4]> 
#>  8  2021     1 108    <int [5]>  <int [7]> 
#>  9  2021     1 109    <int [6]>  <int [4]> 
#> 10  2021     1 110    <int [6]>  <int [9]> 
#> # ℹ 36 more rows
#> 
#> $pop
#> # A tibble: 38 × 4
#>     year   sex agegrp     rks
#>    <int> <int> <fct>    <dbl>
#>  1  2021     1 0 岁      4726
#>  2  2021     1 1-4 岁   36862
#>  3  2021     1 5-9 岁   50162
#>  4  2021     1 10-14 岁 41015
#>  5  2021     1 15-19 岁 41043
#>  6  2021     1 20-24 岁 40989
#>  7  2021     1 25-29 岁 59463
#>  8  2021     1 30-34 岁 49426
#>  9  2021     1 35-39 岁 39549
#> 10  2021     1 40-44 岁 40925
#> # ℹ 28 more rows
#> 
#> attr(,"class")
#> [1] "fbswicd" "list"   
#> 
#> $`411721`
#> $areacode
#> [1] "411721"
#> 
#> $fbswicd
#> # A tibble: 1,596 × 11
#>     year   sex cancer agegrp     fbs   sws    mv   dco m8000    ub   sub
#>    <int> <int> <chr>  <fct>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2016     0 101    0 岁         0     0     0     0     0     0     0
#>  2  2016     0 101    1-4 岁       0     0     0     0     0     0     0
#>  3  2016     0 101    5-9 岁       0     0     0     0     0     0     0
#>  4  2016     0 101    10-14 岁     0     0     0     0     0     0     0
#>  5  2016     0 101    15-19 岁     0     0     0     0     0     0     0
#>  6  2016     0 101    20-24 岁     1     0     1     0     0     0     1
#>  7  2016     0 101    25-29 岁     0     0     0     0     0     0     0
#>  8  2016     0 101    30-34 岁     0     0     0     0     0     0     0
#>  9  2016     0 101    35-39 岁     0     0     0     0     0     0     0
#> 10  2016     0 101    40-44 岁     0     0     0     0     0     0     0
#> # ℹ 1,586 more rows
#> 
#> $sitemorp
#> # A tibble: 46 × 5
#>     year   sex cancer site      morp     
#>    <int> <int> <chr>  <list>    <list>   
#>  1  2016     1 101    <int [8]> <int [4]>
#>  2  2016     1 102    <int [1]> <int [1]>
#>  3  2016     1 103    <int [7]> <int [8]>
#>  4  2016     1 104    <int [5]> <int [6]>
#>  5  2016     1 105    <int [9]> <int [6]>
#>  6  2016     1 106    <int [4]> <int [6]>
#>  7  2016     1 107    <int [3]> <int [5]>
#>  8  2016     1 108    <int [3]> <int [5]>
#>  9  2016     1 109    <int [2]> <int [1]>
#> 10  2016     1 110    <int [6]> <int [5]>
#> # ℹ 36 more rows
#> 
#> $pop
#> # A tibble: 38 × 4
#>     year   sex agegrp     rks
#>    <int> <int> <fct>    <dbl>
#>  1  2016     1 0 岁      5053
#>  2  2016     1 1-4 岁   17743
#>  3  2016     1 5-9 岁   25541
#>  4  2016     1 10-14 岁 32509
#>  5  2016     1 15-19 岁 30530
#>  6  2016     1 20-24 岁 34806
#>  7  2016     1 25-29 岁 36846
#>  8  2016     1 30-34 岁 38691
#>  9  2016     1 35-39 岁 40056
#> 10  2016     1 40-44 岁 39252
#> # ℹ 28 more rows
#> 
#> attr(,"class")
#> [1] "fbswicd" "list"   
#> 
#> attr(,"class")
#> [1] "fbswicds" "list"