Cut age into groups.

Cut age into groups.

Usage

cutage(
  x,
  method = "distance",
  length = 5,
  maxage = 85,
  sep_zero = TRUE,
  breaks = c(seq(0, 85, 5)),
  labels = NULL,
  lang = "cn",
  label_tail = NULL,
  right = FALSE
)

Arguments

x

Vector contains the ages.

method

The method to use for age grouping. Options are "interval", "distance", or "quantile".

length

The length of intervals for age grouping when method is set to "distance".

maxage

The maximum age for age grouping.

sep_zero

A logical value indicating whether to include a separate group for age 0.

breaks

Custom breakpoints for the "interval" method.

labels

labels for the levels of the resulting category. By default, labels are constructed using "(a,b]" interval notation. If labels = FALSE, simple integer codes are returned instead of a factor.

lang

Language used for output. Options are 'cn' or 'en'. Default is 'cn.'.

label_tail

A string to be appended at the end of each label.

right

A logical value indicating whether the intervals are right-closed or right-open.

Value

Factor of age groups.

Examples

age <- sample(0:101, 200, replace = TRUE)
agegrp <- cutage(age,
  method = "distance", length = 5,
  maxage = 60, sep_zero = TRUE)
agegrp
#>   [1] 55-59 岁 25-29 岁 45-49 岁 60+ 岁   40-44 岁 10-14 岁 50-54 岁 60+ 岁  
#>   [9] 15-19 岁 30-34 岁 10-14 岁 0 岁     50-54 岁 20-24 岁 60+ 岁   20-24 岁
#>  [17] 60+ 岁   60+ 岁   60+ 岁   1-4 岁   15-19 岁 50-54 岁 60+ 岁   60+ 岁  
#>  [25] 40-44 岁 60+ 岁   0 岁     60+ 岁   0 岁     15-19 岁 60+ 岁   60+ 岁  
#>  [33] 15-19 岁 10-14 岁 10-14 岁 60+ 岁   60+ 岁   15-19 岁 50-54 岁 1-4 岁  
#>  [41] 15-19 岁 35-39 岁 45-49 岁 50-54 岁 55-59 岁 25-29 岁 60+ 岁   60+ 岁  
#>  [49] 60+ 岁   55-59 岁 60+ 岁   30-34 岁 60+ 岁   55-59 岁 1-4 岁   0 岁    
#>  [57] 60+ 岁   1-4 岁   15-19 岁 25-29 岁 5-9 岁   60+ 岁   55-59 岁 10-14 岁
#>  [65] 40-44 岁 5-9 岁   5-9 岁   55-59 岁 60+ 岁   45-49 岁 60+ 岁   15-19 岁
#>  [73] 35-39 岁 25-29 岁 45-49 岁 60+ 岁   35-39 岁 40-44 岁 25-29 岁 35-39 岁
#>  [81] 10-14 岁 15-19 岁 20-24 岁 1-4 岁   60+ 岁   20-24 岁 60+ 岁   60+ 岁  
#>  [89] 40-44 岁 40-44 岁 60+ 岁   20-24 岁 25-29 岁 60+ 岁   5-9 岁   30-34 岁
#>  [97] 60+ 岁   15-19 岁 60+ 岁   60+ 岁   5-9 岁   1-4 岁   20-24 岁 40-44 岁
#> [105] 60+ 岁   15-19 岁 60+ 岁   60+ 岁   60+ 岁   60+ 岁   5-9 岁   15-19 岁
#> [113] 50-54 岁 40-44 岁 30-34 岁 60+ 岁   40-44 岁 60+ 岁   60+ 岁   10-14 岁
#> [121] 5-9 岁   60+ 岁   5-9 岁   60+ 岁   5-9 岁   20-24 岁 10-14 岁 10-14 岁
#> [129] 60+ 岁   5-9 岁   60+ 岁   5-9 岁   1-4 岁   60+ 岁   60+ 岁   60+ 岁  
#> [137] 35-39 岁 15-19 岁 10-14 岁 60+ 岁   60+ 岁   0 岁     55-59 岁 35-39 岁
#> [145] 40-44 岁 40-44 岁 25-29 岁 30-34 岁 15-19 岁 60+ 岁   5-9 岁   10-14 岁
#> [153] 60+ 岁   20-24 岁 60+ 岁   60+ 岁   35-39 岁 60+ 岁   60+ 岁   60+ 岁  
#> [161] 60+ 岁   60+ 岁   1-4 岁   60+ 岁   40-44 岁 1-4 岁   60+ 岁   30-34 岁
#> [169] 55-59 岁 60+ 岁   60+ 岁   25-29 岁 60+ 岁   30-34 岁 60+ 岁   60+ 岁  
#> [177] 45-49 岁 50-54 岁 30-34 岁 50-54 岁 10-14 岁 60+ 岁   30-34 岁 60+ 岁  
#> [185] 55-59 岁 60+ 岁   50-54 岁 45-49 岁 30-34 岁 60+ 岁   15-19 岁 60+ 岁  
#> [193] 60+ 岁   60+ 岁   60+ 岁   45-49 岁 35-39 岁 60+ 岁   0 岁     50-54 岁
#> 14 Levels: 0 岁 1-4 岁 5-9 岁 10-14 岁 15-19 岁 20-24 岁 25-29 岁 ... 60+ 岁