Expand five-year(abridged) life table to one-year(complete) life table.

Expand five-year(abridged) life table to one-year(complete) life table.

Usage

expand_lifetable(lx)

Arguments

lx

A vector contains the total number of survivors lx at age x which was obtained from the abridged life table.

Value

List contains two vectors fitlx and fitmx.

Examples

lx <- c(
  100000, 99498.39, 99294.62, 99173.88, 99047.59, 98840.46,
  98521.16, 98161.25, 97636.99, 96900.13, 95718.96, 93930.91,
  91463.21, 87131.41, 80525.02, 70907.59, 58090.75, 41630.48,
  24019.33
)
lx <- lx / 100000
expand_lifetable(lx)
#> $fitlx
#>   [1] 1.000000000 0.994983900 0.994304537 0.993754859 0.993309597 0.992946200
#>   [7] 0.992644701 0.992387537 0.992158374 0.991946916 0.991738800 0.991559060
#>  [13] 0.991347165 0.991096388 0.990805228 0.990475900 0.990146402 0.989778116
#>  [19] 0.989366890 0.988909653 0.988404600 0.987828760 0.987212916 0.986565920
#>  [25] 0.985896357 0.985211600 0.984552899 0.983875355 0.983167860 0.982417917
#>  [31] 0.981612500 0.980705363 0.979728767 0.978681787 0.977563049 0.976369900
#>  [37] 0.975129994 0.973793402 0.972340482 0.970750272 0.969001300 0.967057696
#>  [43] 0.964918540 0.962569391 0.959997044 0.957189600 0.954059096 0.950697339
#>  [49] 0.947116296 0.943321677 0.939309100 0.935255836 0.930870900 0.926047151
#>  [55] 0.920671732 0.914632100 0.907632896 0.899826798 0.891186182 0.881688768
#>  [61] 0.871314100 0.860165097 0.848051422 0.834905389 0.820659692 0.805250200
#>  [67] 0.788512450 0.770535568 0.751308576 0.730824120 0.709075900 0.686309280
#>  [73] 0.662182445 0.636612889 0.609534445 0.580907500 0.550378334 0.518493866
#>  [79] 0.485392675 0.451254958 0.416304800 0.380810946 0.345085560 0.309480475
#>  [85] 0.274380479 0.240193300 0.207336178 0.176219242 0.147226296 0.120694131
#>  [91] 0.096891961 0.076003005 0.058110507 0.043190385 0.031112291 0.021649946
#>  [97] 0.014500329 0.009309887 0.005704549 0.003319473
#> 
#> $fitmx
#>   [1] 0.0050287229 0.0006830210 0.0005529797 0.0004481607 0.0003659114
#>   [6] 0.0003036867 0.0002591035 0.0002309476 0.0002131515 0.0002098278
#>  [11] 0.0001812540 0.0002137218 0.0002529973 0.0002938187 0.0003324396
#>  [16] 0.0003327221 0.0003720199 0.0004155588 0.0004622586 0.0005108472
#>  [21] 0.0005827656 0.0006236256 0.0006555918 0.0006789103 0.0006947944
#>  [26] 0.0006688121 0.0006884115 0.0007193483 0.0007630738 0.0008201671
#>  [31] 0.0009245562 0.0009963060 0.0010692143 0.0011437613 0.0012212793
#>  [36] 0.0012707215 0.0013716207 0.0014931349 0.0016367844 0.0018032956
#>  [41] 0.0020077953 0.0022144750 0.0024375255 0.0026759527 0.0029287141
#>  [46] 0.0032758761 0.0035298582 0.0037738668 0.0040145440 0.0042627408
#>  [51] 0.0043244915 0.0046995143 0.0051954468 0.0058216052 0.0065816393
#>  [56] 0.0076819105 0.0086376992 0.0096489341 0.0107142412 0.0118365910
#>  [61] 0.0128781878 0.0141830712 0.0156228608 0.0172098923 0.0189554831
#>  [66] 0.0210048414 0.0230623792 0.0252693615 0.0276436252 0.0302102564
#>  [71] 0.0326342016 0.0357872566 0.0393793535 0.0434663010 0.0481039243
#>  [76] 0.0539856150 0.0596777239 0.0659699946 0.0729257066 0.0806148116
#>  [81] 0.0891146367 0.0985106621 0.1088973812 0.1203792500 0.1330717385
#>  [86] 0.1471024914 0.1626126119 0.1797580810 0.1987113256 0.2196629532
#>  [91] 0.2428236683 0.2684263914 0.2967286020 0.3280149273 0.3626000048
#>  [96] 0.4008316468 0.4430943379 0.4898131021 0.5414577764 0.5414577764
#>