Expand five-year(abridged) life table to one-year(complete) life table.

Expand five-year(abridged) life table to one-year(complete) life table.

Usage

expand_lifetable(lx)

Arguments

lx

A vector contains the total number of survivors lx at age x which was obtained from the abridged life table.

Value

List contains two vectors fitlx and fitmx.

Examples

lx <- c(
  100000, 99498.39, 99294.62, 99173.88, 99047.59, 98840.46,
  98521.16, 98161.25, 97636.99, 96900.13, 95718.96, 93930.91,
  91463.21, 87131.41, 80525.02, 70907.59, 58090.75, 41630.48,
  24019.33
)
lx <- lx / 100000
expand_lifetable(lx)
#> $fitlx
#>   [1] 1.000000e+00 9.949839e-06 9.943045e-06 9.937549e-06 9.933096e-06
#>   [6] 9.929462e-06 9.926447e-06 9.923875e-06 9.921584e-06 9.919469e-06
#>  [11] 9.917388e-06 9.915591e-06 9.913472e-06 9.910964e-06 9.908052e-06
#>  [16] 9.904759e-06 9.901464e-06 9.897781e-06 9.893669e-06 9.889097e-06
#>  [21] 9.884046e-06 9.878288e-06 9.872129e-06 9.865659e-06 9.858964e-06
#>  [26] 9.852116e-06 9.845529e-06 9.838754e-06 9.831679e-06 9.824179e-06
#>  [31] 9.816125e-06 9.807054e-06 9.797288e-06 9.786818e-06 9.775630e-06
#>  [36] 9.763699e-06 9.751300e-06 9.737934e-06 9.723405e-06 9.707503e-06
#>  [41] 9.690013e-06 9.670577e-06 9.649185e-06 9.625694e-06 9.599970e-06
#>  [46] 9.571896e-06 9.540591e-06 9.506973e-06 9.471163e-06 9.433217e-06
#>  [51] 9.393091e-06 9.352558e-06 9.308709e-06 9.260472e-06 9.206717e-06
#>  [56] 9.146321e-06 9.076329e-06 8.998268e-06 8.911862e-06 8.816888e-06
#>  [61] 8.713141e-06 8.601651e-06 8.480514e-06 8.349054e-06 8.206597e-06
#>  [66] 8.052502e-06 7.885124e-06 7.705356e-06 7.513086e-06 7.308241e-06
#>  [71] 7.090759e-06 6.863093e-06 6.621824e-06 6.366129e-06 6.095344e-06
#>  [76] 5.809075e-06 5.503783e-06 5.184939e-06 4.853927e-06 4.512550e-06
#>  [81] 4.163048e-06 3.808109e-06 3.450856e-06 3.094805e-06 2.743805e-06
#>  [86] 2.401933e-06 2.073362e-06 1.762192e-06 1.472263e-06 1.206941e-06
#>  [91] 9.689196e-07 7.600301e-07 5.811051e-07 4.319038e-07 3.111229e-07
#>  [96] 2.164995e-07 1.450033e-07 9.309887e-08 5.704549e-08 3.319473e-08
#> 
#> $fitmx
#>   [1] 1.151795e+01 6.830210e-04 5.529797e-04 4.481607e-04 3.659114e-04
#>   [6] 3.036867e-04 2.591035e-04 2.309476e-04 2.131515e-04 2.098278e-04
#>  [11] 1.812540e-04 2.137218e-04 2.529973e-04 2.938187e-04 3.324396e-04
#>  [16] 3.327221e-04 3.720199e-04 4.155588e-04 4.622586e-04 5.108472e-04
#>  [21] 5.827656e-04 6.236256e-04 6.555918e-04 6.789103e-04 6.947944e-04
#>  [26] 6.688121e-04 6.884115e-04 7.193483e-04 7.630738e-04 8.201671e-04
#>  [31] 9.245562e-04 9.963060e-04 1.069214e-03 1.143761e-03 1.221279e-03
#>  [36] 1.270722e-03 1.371621e-03 1.493135e-03 1.636784e-03 1.803296e-03
#>  [41] 2.007795e-03 2.214475e-03 2.437525e-03 2.675953e-03 2.928714e-03
#>  [46] 3.275876e-03 3.529858e-03 3.773867e-03 4.014544e-03 4.262741e-03
#>  [51] 4.324492e-03 4.699514e-03 5.195447e-03 5.821605e-03 6.581639e-03
#>  [56] 7.681911e-03 8.637699e-03 9.648934e-03 1.071424e-02 1.183659e-02
#>  [61] 1.287819e-02 1.418307e-02 1.562286e-02 1.720989e-02 1.895548e-02
#>  [66] 2.100484e-02 2.306238e-02 2.526936e-02 2.764363e-02 3.021026e-02
#>  [71] 3.263420e-02 3.578726e-02 3.937935e-02 4.346630e-02 4.810392e-02
#>  [76] 5.398562e-02 5.967772e-02 6.596999e-02 7.292571e-02 8.061481e-02
#>  [81] 8.911464e-02 9.851066e-02 1.088974e-01 1.203792e-01 1.330717e-01
#>  [86] 1.471025e-01 1.626126e-01 1.797581e-01 1.987113e-01 2.196630e-01
#>  [91] 2.428237e-01 2.684264e-01 2.967286e-01 3.280149e-01 3.626000e-01
#>  [96] 4.008316e-01 4.430943e-01 4.898131e-01 5.414578e-01 5.414578e-01
#>