Calculate life table from mortality rates.

Calculate life table from mortality rates.

lt(mx, sage = 0, agegrp = 5, sex = "male")

Arguments

mx

Death rate at age x.

sage

Start age.

agegrp

Age groups width.

sex

Gender for the life table, it could be "male","female" or "total".

Value

A data frame containing the following values.

age

Start age in each age group.

mx

Death rate at age x.

qx

The probability that an individual of exact age x will die before exact age x+1.

lx

Number of survivors to exact age x. The radix is 1.

dx

The number of deaths between exact age x and x+1.

Lx

Number of years lived between age x and x+1.

Tx

Number of years lived after exact age x.

ex

Remaining life expectancy at exact age x.

Examples

px <- c(
  20005, 86920, 102502, 151494, 182932,
  203107, 240289, 247076, 199665, 163820,
  145382, 86789, 69368, 51207, 39112, 20509,
  12301, 6586, 1909
)
dx <- c(
  156, 58, 47, 49, 48, 68, 120, 162, 160, 294,
  417, 522, 546, 628, 891, 831, 926, 731, 269
)
mx <- dx / px
lt(mx, sage = 0, agegrp = 5, sex = "total")
#>    age           mx          qx        lx          dx       Lx        Tx
#> 1    0 0.0077980505 0.007741928 1.0000000 0.007741928 0.992803 75.704586
#> 2    1 0.0006672803 0.002664799 0.9922581 0.002644169 3.962606 74.711783
#> 3    5 0.0004585276 0.002290118 0.9896139 0.002266333 4.942630 70.749177
#> 4   10 0.0003234452 0.001615971 0.9873476 0.001595525 4.932909 65.806547
#> 5   15 0.0002623926 0.001311137 0.9857520 0.001292456 4.925658 60.873638
#> 6   20 0.0003347989 0.001672650 0.9844596 0.001646657 4.918346 55.947980
#> 7   25 0.0004993986 0.002494004 0.9828129 0.002451139 4.908182 51.029634
#> 8   30 0.0006556687 0.003273193 0.9803618 0.003208913 4.894108 46.121452
#> 9   35 0.0008013422 0.003999020 0.9771529 0.003907654 4.876386 41.227345
#> 10  40 0.0017946527 0.008934780 0.9732452 0.008695732 4.845356 36.350959
#> 11  45 0.0028683056 0.014243477 0.9645495 0.013738538 4.789775 31.505602
#> 12  50 0.0060145871 0.029645010 0.9508110 0.028186800 4.686406 26.715827
#> 13  55 0.0078710645 0.038625662 0.9226242 0.035636969 4.527592 22.029421
#> 14  60 0.0122639483 0.059566493 0.8869872 0.052834716 4.308133 17.501829
#> 15  65 0.0227807323 0.107998953 0.8341525 0.090087593 3.954552 13.193696
#> 16  70 0.0405187966 0.184638766 0.7440649 0.137383221 3.390605  9.239144
#> 17  75 0.0752784326 0.318795874 0.6066817 0.193407609 2.569230  5.848539
#> 18  80 0.1109930155 0.438228382 0.4132740 0.181108418 1.631710  3.279309
#> 19  85 0.1409114720 1.000000000 0.2321656 0.232165631 1.647599  1.647599
#>           ex
#> 1  75.704586
#> 2  75.294709
#> 3  71.491697
#> 4  66.649829
#> 5  61.753499
#> 6  56.831160
#> 7  51.922021
#> 8  47.045338
#> 9  42.191294
#> 10 37.350256
#> 11 32.663541
#> 12 28.097938
#> 13 23.876917
#> 14 19.731772
#> 15 15.816888
#> 16 12.417121
#> 17  9.640211
#> 18  7.934951
#> 19  7.096654